requirements.txt
fastapi
uvicorn
openai
requests
boto3
python-dotenv
ruff
black
mypy
